Cast size_t to unsigned long for printf to deal with (32-bit) systems where size_t is typedef int.
It's either this or use the C99 %zu but while we can probably
safely assume C99 compilers, I am not at all sure that that's the
case for stdio libraries and this solution is fairly low-impact.


git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@6535 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Common/Src/SymbolDB.cpp
@@ -28,7 +28,7 @@ void SymbolDB::List()
 			iter->second.numCalls);
 	}
 	INFO_LOG(OSHLE, "%lu functions known in this program above.",
-		functions.size());
+		(unsigned long)functions.size());
 }
 
 void SymbolDB::Clear(const char *prefix)
